+++ backend/src/main/java/aimo/backend/domains/security/filter/jwtFilter/JwtAuthenticationFilter.java
@@ -1,9 +1,8 @@
-package aimo.backend.domains.auth.security.jwtFilter;
+package aimo.backend.domains.security.filter.jwtFilter;
 
 import static aimo.backend.common.exception.ErrorCode.*;
 
 import aimo.backend.common.exception.ApiException;
-import aimo.backend.common.exception.ErrorCode;
 import aimo.backend.common.properties.SecurityProperties;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
